:recycle: clean exec code
